feat(prd-6): ci_fix run type + verification (M4)
Server foundation for the CI-fix agent (PRD #6 Phase 2 server half).

- Migration 00043: runs gains kind (issue|ci_fix), nullable issue_iid + a
  runs_kind_shape CHECK pinning the per-kind shape, pipeline_id/pipeline_ref/
  failure_snapshot/fix_verdict. The one-active-per-issue index is re-scoped to
  kind='issue'; a new uq_runs_one_active_ci_fix indexes one active fix per ref.
- POST /repos/{id}/ci-fix-runs: re-checks the pipeline cache shows the ref
  failed, snapshots the failed pipeline (<=CI_FIX_MAX_JOBS jobs x
  CI_FIX_LOG_TAIL_BYTES tails, driver PAT-scrub + a glpat-* scrub), then
  workersvc.CreateCIFixRun enforces the cross-kind same-branch exclusion (both
  create paths) + the one-active-fix index.
- Claim payload gains kind + a pipeline snapshot (issue_iid now nullable);
  the state report gains an outbound fix_verdict ('not_code') threaded into
  SetRunCompleted. Both pinned by a cross-side golden wire-contract test
  (testdata/claim_ci_fix_wire.json).
- Verification loop: the pipeline sync stamps a fix run verified/fix_failed
  from its post-fix pipeline, keyed on runs.branch with the observed-id >
  snapshot-id guard (FindCIFixStampTarget + StampFixVerdict).
- Nullable issue_iid cascade handled across lifecycle (skips ci_fix runs),
  board card queries (issue runs only), mr-watch, autopilot, DTOs.
- Tests: workersvc preconditions + fix_verdict + ci_fix claim golden;
  forgesvc verification stamping (verified/fix_failed/running); a live-DB
  store integration test covering the CHECK, both partial indexes, cross-kind
  counts, and stamp-target selection (migration applies to v43).

+++ api/internal/handler/board_latestrun_test.go
@@ -95,9 +95,9 @@ func TestAssembleCards(t *testing.T) {
 	// Deliberately NOT in issue order (20 before 10): a correct assembly keys by
 	// issue_iid, so a positional/cross-keying bug would surface here.
 	runRows := []store.ListLatestRunsForRepoRow{
-		{IssueIid: 20, ID: run20, UserID: other, Status: "completed", MrIid: i8(5),
+		{IssueIid: i8(20), ID: run20, UserID: other, Status: "completed", MrIid: i8(5),
 			OwnerName: nullTxt(), OwnerEmail: txt("o@example.com"), RunCount: 2, CreatedAt: tstamp(now), UpdatedAt: tstamp(now)},
-		{IssueIid: 10, ID: run10, UserID: viewer, Status: "running",
+		{IssueIid: i8(10), ID: run10, UserID: viewer, Status: "running",
 			OwnerName: txt("Vlad"), WorkerName: txt("laptop"), RunCount: 1, CreatedAt: tstamp(now), UpdatedAt: tstamp(now)},
 	}
 	position := map[string]int{"In Progress": 0}
